[Improve] moduleに切り出してリファクタリング
