+++ app/workers/add_tweets_worker.rb
@@ -1,4 +1,5 @@
 class AddTweetsWorker
+  include LoggerHelper
   include Sidekiq::Worker
   sidekiq_options backtrace: 20
 
@@ -16,21 +17,5 @@ class AddTweetsWorker
   rescue Twitter::Error::TooManyRequests => e
     logger.error('Twitter::Error::TooManyRequestsなので15分後に再試行します。')
     raise e
-  end
-
-  private
-
-  attr_reader :logger
-
-  def initialize
-    set_logger
-  end
-
-  def set_logger
-    @logger = ActiveSupport::Logger.new(Rails.root.join('log/batch.log'))
-    console = ActiveSupport::Logger.new(STDOUT) # この二行がないとlogger.infoが反映されない
-    @logger.extend ActiveSupport::Logger.broadcast(console)
-    @logger.level = Rails.logger.level # この二行でRails.loggerで出力したものを拾う
-    Rails.logger = @logger
   end
 end
